+++ tests/testthat/test-dtt.dispRity.R
@@ -50,6 +50,8 @@ test_that("dispRity and dtt give the same results", {
 
     ## Calculate the disparity of the dataset using dtt::geiger
     set.seed(1)
+    warn <- capture_warning(dtt(phy = geiger_data$phy, data = geiger_data$dat, nsim = 100, plot = FALSE, calculateMDIp = TRUE))
+    expect_equal(as.character(warn), "simpleWarning in treedata(phy, data): The following tips were not found in 'data' and were dropped from 'phy':\n\tolivacea\n")
     expect_warning(geiger_dtt <- dtt(phy = geiger_data$phy, data = geiger_data$dat, nsim = 100, plot = FALSE, calculateMDIp = TRUE))
 
     ## The average squared pairwise distance metric (used in geiger::dtt)
@@ -59,9 +61,15 @@ test_that("dispRity and dtt give the same results", {
     set.seed(1)
     expect_warning(dispRity_dtt <- dtt.dispRity(data = geiger_data$dat, metric = average.sq, tree = geiger_data$phy, nsim = 100))
     plot(dispRity_dtt)
+    expect_error(dispRity_dtt <- dtt.dispRity(data = geiger_data$dat, metric = var, tree = geiger_data$phy, nsim = 100))
      
     ## Same output
     expect_equal(length(geiger_dtt), length(dispRity_dtt))
     expect_equal(unlist(lapply(geiger_dtt[-5], length)), unlist(lapply(dispRity_dtt[-5], length)))
 
+    ## dispRity object
+    data(disparity)
+    data(BeckLee_tree)
+    expect_warning(dtt_data <- dtt.dispRity(data = disparity, metric = average.sq, tree = BeckLee_tree, nsim = 5))
+    expect_is(dtt_data, "dispRity", "dtt")
 })